+++ compiler-plugin-tests/src/test/java/io/ballerina/stdlib/kafka/plugin/CodeSnippetGenerationCodeActionTest.java
@@ -22,14 +22,16 @@ import io.ballerina.projects.plugins.codeaction.CodeActionArgument;
 import io.ballerina.projects.plugins.codeaction.CodeActionInfo;
 import io.ballerina.tools.text.LinePosition;
 import io.ballerina.tools.text.LineRange;
-import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 
 import java.io.IOException;
 import java.nio.file.Path;
 import java.util.List;
 
+import static io.ballerina.stdlib.kafka.plugin.CompilerPluginTestUtils.BALLERINA_SOURCES;
+import static io.ballerina.stdlib.kafka.plugin.CompilerPluginTestUtils.EXPECTED_SOURCES;
 import static io.ballerina.stdlib.kafka.plugin.KafkaCodeTemplate.NODE_LOCATION;
+import static io.ballerina.stdlib.kafka.plugin.CompilerPluginTestUtils.RESOURCE_DIRECTORY;
 
 /**
  * A class for testing code actions.
@@ -39,10 +41,10 @@ public class CodeSnippetGenerationCodeActionTest extends AbstractCodeActionTest 
     @Test
     public void testEmptyServiceCodeAction()
             throws IOException {
-        Path filePath = RESOURCE_PATH.resolve("ballerina_sources")
+        Path filePath = RESOURCE_DIRECTORY.resolve(BALLERINA_SOURCES)
                 .resolve("snippet_gen_service_1")
                 .resolve("service.bal");
-        Path resultPath = RESOURCE_PATH.resolve("expected_sources")
+        Path resultPath = RESOURCE_DIRECTORY.resolve(EXPECTED_SOURCES)
                 .resolve("service_1")
                 .resolve("result.bal");
         performTest(filePath, LinePosition.from(2, 0),
@@ -52,10 +54,10 @@ public class CodeSnippetGenerationCodeActionTest extends AbstractCodeActionTest 
     @Test
     public void testServiceWithVariablesCodeAction()
             throws IOException {
-        Path filePath = RESOURCE_PATH.resolve("ballerina_sources")
+        Path filePath = RESOURCE_DIRECTORY.resolve(BALLERINA_SOURCES)
                 .resolve("snippet_gen_service_2")
                 .resolve("service.bal");
-        Path resultPath = RESOURCE_PATH.resolve("expected_sources")
+        Path resultPath = RESOURCE_DIRECTORY.resolve(EXPECTED_SOURCES)
                 .resolve("service_2")
                 .resolve("result.bal");
         performTest(filePath, LinePosition.from(2, 0),
